fix(expert-prediction): cast the ranker's bf16 pre_mixer input to float32
pre_mixer is captured/stored in the model's bf16 hidden dtype; the ranker's
nn.Linear weights are float32, so the matmul raised a dtype mismatch. Cast
at the module boundary, matching OrdinalCDF's existing x.float() handling.

+++ python/sglang/srt/layers/moe/expert_prediction/training/apex.py
@@ -23,7 +23,7 @@ class Ranker(nn.Module):
         self.linear = nn.Linear(hidden_size, num_experts, bias=True)
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
-        return self.linear(x)
+        return self.linear(x.float())
 
 
 def teacher_probabilities(router_input: torch.Tensor, gate_weight: torch.Tensor) -> torch.Tensor:
